+++ src/auth/pages/changepassword/ChangePassword.css
@@ -1,12 +1,4 @@
 
-.page-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 24px 48px;
-  width: 100%;
-  box-sizing: border-box;
-}
 
 .logo {
   font-weight: 700;
@@ -14,22 +6,24 @@
   color: #433f9b;
 }
 
-/* Container de página para centralizar o card */
 .change-password-page {
   display: flex;
   flex-direction: column;
-  justify-content: center;
+  justify-content: space-between;
   align-items: center;
   min-height: 100vh;
-  background-color: #EDE9FB; /* Cor de fundo suave para área logada */
+  height: 100%;
+  background-color: #EDE9FB; 
   font-family: 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
   
 }
 
+
 .change-password-container {
   width: 100%;
-  max-width: 400px;
+  max-width: 30rem;
   padding: 20px;
+  
 }
 
 .card-header {
@@ -53,7 +47,6 @@
   cursor: pointer;
 }
 
-/* Estilo do Card principal */
 .change-password-card {
   background: #ffffff;
   padding: 2rem;
@@ -77,7 +70,6 @@
   margin: 0;
 }
 
-/* Grupos do formulário */
 .form-group {
   margin-bottom: 1.2rem;
   display: flex;
@@ -93,7 +85,6 @@
   margin-bottom: 0.5rem;
 }
 
-/* Wrapper para posicionar o botão de olho dentro do input */
 .input-wrapper {
   position: relative;
   display: flex;
@@ -102,7 +93,7 @@
 
 .input-wrapper input {
   width: 100%;
-  padding: 12px 45px 12px 12px; /* Espaço extra à direita para o ícone */
+  padding: 12px 45px 12px 12px; 
   border: 1px solid #ddd;
   border-radius: 8px;
   font-size: 1rem;
@@ -119,7 +110,6 @@
   border-color: #dc3545;
 }
 
-/* Botão de alternar visibilidade (Olho) */
 .toggle-password {
   position: absolute;
   right: 12px;
@@ -138,7 +128,6 @@
   color: #333;
 }
 
-/* Textos de apoio e erro */
 .helper-text {
   font-size: 0.8rem;
   color: #777;
@@ -152,14 +141,13 @@
   font-weight: 500;
 }
 
-/* Botão de Envio */
 .submit-btn {
   width: 100%;
   padding: 16px;
   background-color: #433f9b;
   color: white;
   border: none;
-  border-radius: 999px; /* Pill shape */
+  border-radius: 999px; 
   font-size: 18px;
   font-weight: 600;
   cursor: pointer;
@@ -183,7 +171,6 @@
 }
 
 
-/* Rodapé do card */
 .card-footer {
   margin-top: 1.5rem;
   text-align: center;
@@ -203,7 +190,6 @@
   text-decoration: underline;
 }
 
-/* --- Rodapé da Página --- */
 .page-footer {
   display: flex;
   justify-content: space-between;
