+++ src/auth/pages/changepassword/ChangePassword.jsx
@@ -1,9 +1,11 @@
 import React, { useState, useEffect } from "react";
 // import { FaEye, FaEyeSlash, FaMoon } from "react-icons/fa";
-import "./alterarsenha.css";
+import "./ChangePassword.css";
+import Header from "../../../global/components/Header/Header";
+import Footer from "../../../global/components/Footer/Footer";
 import { FaArrowLeft } from "react-icons/fa";
 
-const AlterarSenha = () => {
+const ChangePassword = () => {
   // Estados para os valores dos inputs
   const [formData, setFormData] = useState({
     currentPassword: "",
@@ -11,33 +13,26 @@ const AlterarSenha = () => {
     confirmPassword: "",
   });
 
-  // Estados para visibilidade das senhas (individual para cada campo)
   const [showPasswords, setShowPasswords] = useState({
     current: false,
     new: false,
     confirm: false,
   });
 
-  // Estado para mensagens de erro
   const [errors, setErrors] = useState({
     mismatch: false,
     sameAsOld: false,
   });
 
-  // Estado para controlar se o botão está habilitado
   const [isButtonDisabled, setIsButtonDisabled] = useState(true);
 
-  // Efeito para validar o formulário sempre que os dados mudarem
   useEffect(() => {
     const { currentPassword, newPassword, confirmPassword } = formData;
 
-    // Validação: campos preenchidos
     const allFilled = currentPassword && newPassword && confirmPassword;
 
-    // Validação: senhas coincidem
     const mismatch = newPassword !== confirmPassword && confirmPassword !== "";
 
-    // Validação: nova senha igual à atual
     const sameAsOld = newPassword === currentPassword && newPassword !== "";
 
     setErrors({
@@ -45,12 +40,10 @@ const AlterarSenha = () => {
       sameAsOld,
     });
 
-    // O botão só habilita se tudo estiver preenchido, coincidirem e a nova for diferente da antiga
     const canSubmit = allFilled && !mismatch && !sameAsOld;
     setIsButtonDisabled(!canSubmit);
   }, [formData]);
 
-  // Manipulador de mudança nos inputs
   const handleChange = (e) => {
     const { name, value } = e.target;
     setFormData((prev) => ({
@@ -59,7 +52,6 @@ const AlterarSenha = () => {
     }));
   };
 
-  // Alternar visibilidade da senha
   const toggleVisibility = (field) => {
     setShowPasswords((prev) => ({
       ...prev,
@@ -70,10 +62,8 @@ const AlterarSenha = () => {
   const handleSubmit = (e) => {
     e.preventDefault();
     alert("Senha alterada com sucesso! (Simulação)");
-    // Aqui viria a integração com a API
   };
 
-  // Componente interno para os ícones de Olho (SVG)
   const EyeIcon = ({ visible }) => (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -102,6 +92,7 @@ const AlterarSenha = () => {
 
   return (
     <div className="change-password-page">
+      <Header/>
       <main className="change-password-container">
         <section className="change-password-card">
           <div className="card-header">
@@ -144,7 +135,7 @@ const AlterarSenha = () => {
               </div>
             </div>
 
-            {/* Nova Senha */}
+            {}
             <div className="form-group">
               <label htmlFor="newPassword" style={{ textAlign: "left" }}>
                 Nova senha
@@ -181,7 +172,7 @@ const AlterarSenha = () => {
               )}
             </div>
 
-            {/* Confirmar Nova Senha */}
+            {}
             <div className="form-group">
               <label htmlFor="confirmPassword" style={{ textAlign: "left" }}>
                 Confirme a nova senha
@@ -229,8 +220,9 @@ const AlterarSenha = () => {
           </div>
         </section>
       </main>
+      <Footer/>
     </div>
   );
 };
 
-export default AlterarSenha;
+export default ChangePassword;
